requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=1.10.0,<2.0.0
redis>=5.0.0
structlog>=23.2.0
python-dotenv>=1.0.0
tenacity>=8.2.3
prometheus-client>=0.19.0
httpx>=0.25.2 